% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_smooth.R
\name{exp_smooth}
\alias{exp_smooth}
\title{Smooth a time series}
\usage{
exp_smooth(y, a)
}
\arguments{
\item{y}{Time series to be smoothed}

\item{a}{Smooth parameter}
}
\value{
A time series smoothed from \code{y}
}
\description{
A function to smooth a time series
}
\examples{
set.seed(10)
y <- cumsum(rnorm(1e5))
ys <- exp_smooth(y, 0.8)
plot(y, col = "grey80")
lines(ys, col = "red")

}
