% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mutual_information}
\alias{mutual_information}
\title{Calculate the mutual information}
\usage{
mutual_information(joint_dist)
}
\arguments{
\item{joint_dist}{A numeric matrix}
}
\value{
The mutual information
}
\description{
Calculate the mutual information
}
